pyproject.toml
[tool.poetry]
name = "monarch-agent"
version = "0.1.0"
description = "LLM retrieval augmented generation agent for Monarch."
authors = ["Shawn O'Neil <oneilsh@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
agent-smith-ai = "^0.10.2"
streamlit = "^1.26.0"
python-dotenv = "^1.0.0"
streamlit-cookies-manager = {git = "https://github.com/oneilsh/streamlit-cookies-manager.git"}
streamlit-ws-localstorage = "^1.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
